analyses/human-speech-filtering/app/requirements.txt
absl-py==0.12.0; 
alembic==1.6.5; 
appdirs==1.4.4; 
attrs==21.2.0; 
audioread==2.1.9; 
cachecontrol==0.12.11; 
cachetools==4.2.2; 
certifi==2021.5.30; 
cffi==1.14.5; 
chardet==4.0.0; 
click==7.1.2; 
cliff==3.8.0; 
cmaes==0.8.2; 
cmd2==2.1.2; 
colorama==0.4.4; 
colorlog==5.0.1; 
cycler==0.10.0; 
decorator==5.0.9; 
docopt==0.6.2; 
filelock==3.0.12; 
firebase-admin==5.2.0; 
fs.smbfs==1.0.2; 
fs==2.4.13; 
google-api-core==2.7.3; 
google-api-python-client==2.47.0; 
google-auth-httplib2==0.1.0; 
google-auth-oauthlib==0.4.4; 
google-auth==1.30.1; 
google-cloud-core==2.3.0; 
google-cloud-firestore==2.4.0; 
google-cloud-pubsub==2.12.1; 
google-cloud-storage==2.2.1; 
google-crc32c==1.3.0; 
google-resumable-media==2.3.2; 
googleapis-common-protos==1.56.1; 
greenlet==1.1.0; 
grpc-google-iam-v1==0.12.4; 
grpcio-status==1.46.1; 
grpcio==1.46.1; 
httplib2==0.20.4; 
idna==2.10; 
iprogress==0.4; 
joblib==1.0.1; 
kiwisolver==1.3.1; 
librosa==0.8.1; 
llvmlite; 
mako==1.1.4; 
markdown==3.3.4; 
markupsafe==2.0.1; 
matplotlib==3.4.2; 
mpmath==1.2.1; 
msgpack==1.0.3; 
numba==0.50.1;
numpy; 
oauthlib==3.1.1; 
optuna==2.8.0; 
packaging==20.9; 
pandas==1.1.5; 
pbr==5.6.0; 
pescador==2.1.0; 
pillow==8.2.0; 
pooch==1.3.0; 
prettytable==2.1.0; 
proto-plus==1.19.6; 
protobuf==3.17.1; 
py==1.10.0; 
pyannote.audio==1.1.1; 
pyannote.core==4.1; 
pyannote.database==4.1; 
pyannote.metrics==3.0.1; 
pyannote.pipeline==1.5.2; 
pyasn1-modules==0.2.8; 
pyasn1==0.4.8; 
pycparser==2.20; 
pydub==0.25.1; 
pyparsing==2.4.7; 
pyperclip==1.8.2; 
pyreadline3==3.3; 
pysmb==1.2.7; 
python-dateutil==2.8.1; 
python-editor==1.0.4; 
pytz==2021.1; 
pyyaml==5.4.1; 
pyzmq==22.1.0; 
requests-oauthlib==1.3.0; 
requests==2.25.1; 
resampy==0.2.2; 
rsa==4.7.2; 
scikit-learn==0.24.2; 
scipy==1.5.4; 
seaborn==0.11.2; 
shellingham==1.4.0; 
simplejson==3.17.3; 
six==1.16.0; 
sklearn==0.0; 
sortedcollections==2.1.0; 
sortedcontainers==2.4.0; 
soundfile==0.10.3.post1; 
sqlalchemy==1.4.21; 
stevedore==3.3.0; 
sympy==1.8; 
tabulate==0.8.9; 
tensorboard-data-server==0.6.1; 
tensorboard-plugin-wit==1.8.0; 
tensorboard==2.5.0; 
threadpoolctl==2.1.0; 
torch==1.9.0;
tqdm==4.61.2; 
typer==0.4.1; 
typing-extensions==3.10.0.0; 
uritemplate==4.1.1; 
urllib3==1.26.5; 
wcwidth==0.2.5; 
werkzeug==2.0.1; 